+++ src/main/java/com/atlauncher/constants/Constants.java
@@ -83,6 +83,7 @@ public class Constants {
     public static final int CURSEFORGE_FABRIC_MODLOADER_ID = 4;
     public static final int CURSEFORGE_PAGINATION_SIZE = 20;
     public static final int CURSEFORGE_FABRIC_MOD_ID = 306612;
+    public static final int CURSEFORGE_LEGACY_FABRIC_MOD_ID = 400281;
     public static final int CURSEFORGE_JUMPLOADER_MOD_ID = 361988;
     public static final int CURSEFORGE_MODS_SECTION_ID = 6;
     public static final int CURSEFORGE_MODPACKS_SECTION_ID = 4471;
@@ -93,6 +94,7 @@ public class Constants {
     public static final String MODRINTH_API_URL = "https://api.modrinth.com/v2";
     public static final String MODRINTH_HOST = "api.modrinth.com";
     public static final String MODRINTH_FABRIC_MOD_ID = "P7dR8mSH";
+    public static final String MODRINTH_LEGACY_FABRIC_MOD_ID = "9CJED7xi";
     public static final String MODRINTH_QSL_MOD_ID = "qvIfYCYJ";
     public static final int MODRINTH_PAGINATION_SIZE = 20;
 
@@ -116,6 +118,10 @@ public class Constants {
     // Fabric domains, endpoints, etc
     public static final String FABRIC_MAVEN = "https://maven.fabricmc.net/";
     public static final String FABRIC_HOST = "maven.fabricmc.net";
+
+    // Legacy Fabric domains, endpoints, etc
+    public static final String LEGACY_FABRIC_MAVEN = "https://maven.legacyfabric.net/";
+    public static final String LEGACY_FABRIC_HOST = "maven.legacyfabric.net";
 
     // Quilt domains, endpoints, etc
     public static final String QUILT_MAVEN = "https://maven.quiltmc.org/repository/release/";