+++ src/main/java/com/atlauncher/gui/dialogs/CurseForgeProjectFileSelectorDialog.java
@@ -247,6 +247,14 @@ public class CurseForgeProjectFileSelectorDialog extends JDialog {
                                     return true;
                                 }
 
+                                // don't show CurseForge dependency when grabbed from Modrinth
+                                if (dependency.modId == Constants.CURSEFORGE_LEGACY_FABRIC_MOD_ID
+                                        && installedMod.isFromModrinth()
+                                        && installedMod.modrinthProject.id
+                                                .equals(Constants.MODRINTH_LEGACY_FABRIC_MOD_ID)) {
+                                    return true;
+                                }
+
                                 return installedMod.isFromCurseForge()
                                         && installedMod.getCurseForgeModId() == dependency.modId;
                             }))
@@ -307,7 +315,7 @@ public class CurseForgeProjectFileSelectorDialog extends JDialog {
             // filter out files not for our loader
             curseForgeFilesStream = curseForgeFilesStream.filter(cf -> {
                 if (cf.gameVersions.contains("Fabric") && loaderVersion != null
-                        && (loaderVersion.isFabric() || loaderVersion.isQuilt())) {
+                        && (loaderVersion.isFabric() || loaderVersion.isLegacyFabric() || loaderVersion.isQuilt())) {
                     return true;
                 }